requirements.txt
# Core
streamlit>=1.29.0
python-dotenv>=1.0.0

# Hugging Face & ML
transformers>=4.36.0
torch>=2.1.0
sentencepiece>=0.1.99
accelerate>=0.25.0

# Korean NLP
konlpy>=0.6.0
keybert>=0.8.3
sentence-transformers>=2.2.2

# Web Scraping
newspaper3k>=0.2.8
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Visualization
plotly>=5.18.0
streamlit-extras>=0.3.0

# Utilities
validators>=0.22.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
